Add an extension to get desktop display id and window id for screen sharing. & Add duo-process support for VideoSurface.

+++ Projects/Scripts/Scripts/IAgoraRtcEngine.cs
@@ -10,7 +10,7 @@ using System;
 
 namespace agora_gaming_rtc
 {
-    using view_t = IntPtr;
+    using view_t = UInt64;
 
     public abstract class IAgoraRtcEngine : IRtcEngine
     {
